feat(bookmark page): add skeleton
