+++ src/app/sn/bookmarks/SavedRecipes/style.module.scss
@@ -22,3 +22,9 @@
     grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
   }
 }
+
+// skeleton
+.skeleton__title {
+  width: 100px;
+  height: 20px;
+}
